requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
email-validator
argon2-cffi
pytest
jinja2
python-multipart
python-jose[cryptography]
algoliasearch>=3.0.0
pydantic-settings